pyproject.toml
[project]
name = "fastapi-financial-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "bcrypt>=4.3.0",
  "fastapi",
  "pymongo>=4.13.2",
  "python-dotenv>=1.1.1",
  "uvicorn",
]
